backend/pyproject.toml
[project]
name = "fastapi-tamplate"
version = "0.1.0"
description = "A Fullstack-FastApi template with all boilerplate Code and Configuration, including : 1. DataBase & ORM (SQLMODEL) 2. API (FASTAPI) 3. TypeSafety & MODEL (PYDANTIC) 4. Authentication (pyjwt)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.116.1",
    "pyjwt>=2.10.1",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
]
